+++ cbits/proteaaudio_binding.h
@@ -1,8 +1,9 @@
 #ifdef __cplusplus
-extern "C" { 
+extern "C" {
 #endif
 
 typedef int sample_t;
+typedef int sound_t;
 
 int initAudio(int nTracks, int frequency, int chunkSize);
 void finishAudio();
@@ -14,10 +15,10 @@ sample_t _sampleFromMemoryOgg(char *data, int size, float volume);
 sample_t sampleFromFile(char* filename, float volume);
 int soundActive();
 void soundStopAll();
-void soundLoop(sample_t sample, float volumeL, float volumeR, float disparity, float pitch);
-void soundPlay(sample_t sample, float volumeL, float volumeR, float disparity, float pitch);
-int soundUpdate(sample_t sample, float volumeL, float volumeR, float disparity, float pitch);
-int soundStop(sample_t sample);
+sound_t soundLoop(sample_t sample, float volumeL, float volumeR, float disparity, float pitch);
+sound_t soundPlay(sample_t sample, float volumeL, float volumeR, float disparity, float pitch);
+int soundUpdate(sound_t sound, float volumeL, float volumeR, float disparity, float pitch);
+int soundStop(sound_t sound);
 #ifdef __cplusplus
 }
 #endif