+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.8.0
+- fix: distinct data and playback handles
+- introduce Sound handle type, an abstraction for playback audio track
+- soundLoop, soundPlay return a Sound handle
+- soundUpdate, soundStop take a Sound handle parameter
+
 # 0.7.1.0
 - add sampleFromMemoryPcm
 - update stb_vorbis to v1.14